requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-ckeditor==4.0.2
django-preventconcurrentlogins==0.8.2
django-richtextfield==1.6
djangorestframework==3.12.1
fcm-django==0.3.4
idna==2.10
Pillow==7.2.0
psycopg2-binary==2.8.6
pyfcm==1.4.7
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
